fix(cuda): nll_loss read float targets as int* — the #775 fix never reached CUDA
TensorNLLLoss[4,8] returned exactly 0 on the GPU:
    forward parity  maxAbs 4.102E+00 @[0]  a=4.1015635  b=0

The kernel declared its targets parameter as `const int*` while the buffer holds
class indices as float VALUES (3.0f == class 3), matching CpuEngine which does
(int)targetValue. Reading it as int* reinterprets the float BIT PATTERN
(1.0f -> 1065353216), which is always >= numClasses, so every element took the
out-of-range else branch and the whole loss came back zero.

This is the SAME partial rollout as the SELU >= defect fixed earlier in this
branch, and the same issue number. #775 corrected nll_loss on OpenCL — whose
kernel carries a comment spelling out this exact bug, including the
"1.0f -> 1065353216" example — and on Metal, but CUDA was never updated. The
registry case even documents it as "FIXED (#775)" while the test kept failing.
HIP has no nll_loss kernel, so it falls back to CPU and is unaffected.

Fixed by taking `const float*` and truncating with (int)targets[b], matching
CpuEngine and the OpenCL/Metal kernels.

Measured:
    OpParityTests   12 failed / 1066 passed -> 10 failed / 1068 passed
    TensorNLLLoss[4,8] -> PASS

Stability run BEFORE writing these numbers:
    TapeGradientParityTests  1 / 1 / 0 failures over three runs — the documented
      intermittent driver artifact, unchanged by this commit
    GpuResidencyProbeTests   1 failed / 10 passed, 8 ops below floor (unchanged)

Remaining forward-parity failures: CrossEntropyBackward, GLU, PositionalEncoding,
TensorEq, TensorIsClose.

+++ src/AiDotNet.Tensors/Engines/DirectGpu/CUDA/Kernels/CudaLossForwardKernels.cs
@@ -206,14 +206,23 @@ extern ""C"" __global__ __launch_bounds__(256) void bce_with_logits_loss(
 
 extern ""C"" __global__ __launch_bounds__(256) void nll_loss(
     const float* __restrict__ logProbs,
-    const int* __restrict__ targets,
+    const float* __restrict__ targets,
     float* __restrict__ loss,
     int batchSize, int numClasses)
 {
     int b = blockIdx.x * blockDim.x + threadIdx.x;
     if (b >= batchSize) return;
 
-    int targetClass = targets[b];
+    // targets holds class indices as float VALUES (3.0f == class 3), matching
+    // CpuEngine.TensorNLLLoss which does (int)targetValue. This kernel declared them as
+    // `const int*` and so reinterpreted the float BIT PATTERN (1.0f -> 1065353216), which is
+    // always out of range, so every element took the else branch and the GPU returned 0.
+    //
+    // Same partial rollout as the SELU >= fix: #775 corrected this on OpenCL (whose kernel
+    // carries the identical explanation) and Metal, but CUDA was never updated — the registry
+    // case still records it as FIXED (#775) while forward parity measured
+    // maxAbs 4.102E+00 @[0] a=4.1015635 b=0, i.e. the CPU loss against a GPU zero.
+    int targetClass = (int)targets[b];
     loss[b] = (targetClass >= 0 && targetClass < numClasses)
         ? -logProbs[b * numClasses + targetClass] : 0.0f;
 }
